wip(rscore): checkpoint production H1 parity work

+++ rscore/crates/process/src/tests/account_input_wire_fixture.rs
@@ -60,7 +60,7 @@ pub(super) fn peer_row(kind: AbiValue) -> AbiValue {
 }
 
 pub(super) fn ack_frame_row() -> AbiValue {
-    peer_row(tuple(vec![AbiValue::Integer(2), ack(), proposal()]))
+    peer_row(tuple(vec![AbiValue::Integer(0), ack(), proposal()]))
 }
 
 pub(super) fn at<'a>(value: &'a AbiValue, path: &[usize]) -> &'a AbiValue {
